requirements.txt
langchain[google-genai]
python-dotenv
langchain
langchain_core
